backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-multipart
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
xgboost
shap
pandas
numpy
reportlab
psutil
watchfiles
python-dotenv
matplotlib
email-validator
scapy
